% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{assignCPAloc}
\alias{assignCPAloc}
\title{assign to single subcellular locaction}
\usage{
assignCPAloc(assignLocProps, cutoff = 0.8, Locations)
}
\arguments{
\item{assignLocProps}{matrix of proportion estimates for each protein}

\item{cutoff}{cutoff for assigning a protein to a location}

\item{Locations}{list of subcellular locations}
}
\value{
protLoc assigned location of protein
}
\description{
Assign proteins to the most prevalent subcellular location 
(meeting a specific cutoff) using CPA estimates,
with minimum value being 0.5.
}
\examples{
data(protNSA_test)
data(markerListJadot)
refLocationProfilesNSA <- locationProfileSetup(profile=protNSA_test,
         markerList=markerListJadot, numDataCols=9)
protCPAfromNSA_test <- fitCPA(profile=protNSA_test,
         refLocationProfiles=refLocationProfilesNSA,
         numDataCols=9)
Locations <- unique(markerListJadot$referenceCompartment)
table(apply(protCPAfromNSA_test[,1:8],1,assignCPAloc,
         cutoff=0.8, Locations=Locations))
}
