% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{fCPAone}
\alias{fCPAone}
\title{CPA for protein i}
\usage{
fCPAone(
  profile,
  refLocationProfiles,
  numDataCols,
  startProps = NULL,
  maxit = 10000,
  ind.vary = NULL,
  minVal = FALSE
)
}
\arguments{
\item{profile}{vector of a specified protein (row name) profile}

\item{refLocationProfiles}{data frame of profiles for the 
reference compartments}

\item{numDataCols}{number of fractions in each profile}

\item{startProps}{starting values for proportional assignments;
set equal if this is null (default)}

\item{maxit}{maximum number of iterations (default is 10000)}

\item{ind.vary}{if not NULL, indexes of proportions allowed to vary with 
others constrained to zero}

\item{minVal}{default is FALSE. If TRUE, return minimum value of 
goodness of fit}
}
\value{
Vector of proportional assignments of
   each protein to compartments. Also returns variables
   'nspectra' and 'npeptides' if they are included in the profile input.
   If 'ind.vary' is specified, only the 
   referenced CPA estimates are returned.
}
\description{
Carry out constrained proportional assignment for protein i;
   service function for fitCPA
}
\examples{
data(protRSA_test)
data(refLocProfRSA)

protCPAfromRSA_i_out1 <- fCPAone(profile=protRSA_test[1,],
                          refLocationProfiles=refLocProfRSA,
                          numDataCols=9, startProps=NULL,
                          maxit=10000,
                         ind.vary=NULL, minVal=FALSE)
round(protCPAfromRSA_i_out1, digits=4)

protCPAfromRSA_i_out1b <- fCPAone(profile=protRSA_test[1,],
          refLocationProfiles=refLocProfRSA,
          numDataCols=9, startProps=NULL,
          maxit=10000,
          ind.vary=c(2,4), minVal=FALSE)
round(protCPAfromRSA_i_out1b, digits=4)
}
