% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCPAsubsets.r
\name{fCPAsubsets}
\alias{fCPAsubsets}
\title{Carry out constrained proportional assignment for protein
 number i in profile
 with constraint being the number of compartments that are allowed to
 vary freely; other compartment proportions are fixed at zero}
\usage{
fCPAsubsets(
  profile,
  refLocationProfiles,
  numDataCols,
  startProps = NULL,
  showProgress = TRUE,
  maxit = 10000,
  nCPAcomparts = 2
)
}
\arguments{
\item{profile}{one-row data frame of a protein name (in the row name) and
relative abundance levels.
          Nspectra and Npep, if present, will be removed}

\item{refLocationProfiles}{A matrix giving the abundance level profiles
of the subcellular locations}

\item{numDataCols}{Number of channels of abundance levels}

\item{startProps}{starting valuese for proportional assignements;
set equal if this is null (default)}

\item{showProgress}{default is T}

\item{maxit}{maximum number of iterations (default is 10000)}

\item{nCPAcomparts}{number of compartments to fit restricted CPA;
remaining proportions are fixed at zero}
}
\value{
data frame with CPA estimates for every combination
       ordered by value, with best fit listed first
}
\description{
Carry out constrained proportional assignment for protein
 number i in profile
 with constraint being the number of compartments that are allowed to
 vary freely; other compartment proportions are fixed at zero
}
\examples{
data(protNSA_test)
data(markerListJadot)
nTestProts <- nrow(protNSA_test)
refLocationProfilesNSA <- locationProfileSetup(profile=protNSA_test,
    markerList=markerListJadot, numDataCols=9)
protCPAfromNSA_test <- fCPAsubsets(profile=protNSA_test[1,],
                         refLocationProfiles=refLocationProfilesNSA,
                         numDataCols=9, startProps=NULL, nCPAcomparts=2)
head(protCPAfromNSA_test)
}
