% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{fitCPA}
\alias{fitCPA}
\title{Carry out constrained proportional assignment on all proteins}
\usage{
fitCPA(
  profile,
  refLocationProfiles,
  numDataCols,
  startProps = NULL,
  maxit = 10000,
  showProgress = TRUE,
  ind.vary = NULL,
  minVal = FALSE
)
}
\arguments{
\item{profile}{a data frame of protein names (row names) and
relative abundance levels.}

\item{refLocationProfiles}{A matrix giving the abundance level
profiles of the subcellular locations}

\item{numDataCols}{Number of channels of abundance levels}

\item{startProps}{starting valuese for proportional assignements;
set equal if this is null (default)}

\item{maxit}{maximum number of iterations (default is 10000)}

\item{showProgress}{print out progress if TRUE, the default}

\item{ind.vary}{if not NULL, indexes of parameters to allow to vary}

\item{minVal}{default is false. If true, return minimum value of
goodness of fit}
}
\value{
assignProbsOut  Data frame of proportional assignments
      of each protein to compartments
}
\description{
Carry out constrained proportional assignment on all proteins
}
\examples{
data(protNSA_test)
data(markerListJadot)
refLocationProfilesNSA <- locationProfileSetup(profile=protNSA_test,
       markerList=markerListJadot, numDataCols=9)
protCPAfromNSA_test <- fitCPA(profile=protNSA_test,
                           refLocationProfiles=refLocationProfilesNSA,
                           numDataCols=9)
}
