% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerProfilePlot.r
\name{markerProfilePlot}
\alias{markerProfilePlot}
\title{Plot profiles of reference proteins}
\usage{
markerProfilePlot(
  refLoc,
  profile,
  markerList,
  refLocationProfiles,
  ylab = "",
  refProtPlot = NULL
)
}
\arguments{
\item{refLoc}{name of the reference subcellular
compartment to plot}

\item{profile}{data frame of specified protein (row name) profiles}

\item{markerList}{list of reference proteins}

\item{refLocationProfiles}{data frame of profiles for the 
reference compartments}

\item{ylab}{label for y=axis, eg, 'NSA'}

\item{refProtPlot}{indices of reference proteins to plot; default is NULL}
}
\value{
Plot of profiles of reference (marker) proteins
}
\description{
This function plots profiles of reference proteins
   and also the average profile for each compartment
}
\examples{
data(protNSA_test)
data(markerListJadot)
data(refLocProfNSA)
markerProfilePlot(refLoc='PM', profile=protNSA_test,
   markerList=markerListJadot,
   refLocationProfiles=refLocProfNSA, ylab='NSA')
}
