% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureHeatMap.r
\name{mixtureHeatMap}
\alias{mixtureHeatMap}
\title{plot mixture of all two compartment profile combinations as panel;
   assumes eight compartments}
\usage{
mixtureHeatMap(Acup, totProt, NstartMaterialFractions = 6, eps = 0.001)
}
\arguments{
\item{Acup}{relative amount of a given cellular
compartment protein that ends up in a given centrifugation fraction}

\item{totProt}{vector of amounts starting material in each fraction}

\item{NstartMaterialFractions}{number of fractions in starting material}

\item{eps}{small positive constant to add before taking a log transformation}
}
\value{
color heat map of errors
}
\description{
plot mixture of all two compartment profile combinations as panel;
   assumes eight compartments
}
\examples{
  # See See Vignette 4 for a full explanation
}
