% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixturePlot.r
\name{mixturePlot}
\alias{mixturePlot}
\title{plot mixture of two compartment profiles}
\usage{
mixturePlot(
  mixProtiProtjCPA,
  NstartMaterialFractions = 6,
  Loc1,
  Loc2,
  increment.prop = 0.1,
  errorReturn = FALSE,
  subTitle = NULL,
  xaxisLab = TRUE,
  yaxisLab = TRUE
)
}
\arguments{
\item{mixProtiProtjCPA}{data frame of CPA estimated proportions
for each mixture}

\item{NstartMaterialFractions}{Number of fractions that comprise
the starting material}

\item{Loc1}{row number of subcellular location 1 of mixture}

\item{Loc2}{row number of subcellular location 2 of mixture}

\item{increment.prop}{increments in proportions of Loc1 vos Loc2;
Default is 0.1}

\item{errorReturn}{Return area of error region if true}

\item{subTitle}{subtitle for plot if present; NULL if not (default)}

\item{xaxisLab}{plot label for x-axis if true}

\item{yaxisLab}{plot label for y-axis if true}
}
\value{
plot of mixture of two compartment profiles
}
\description{
plot mixture of two compartment profiles
}
\examples{
data(protNSA_test)
data(markerListJadot)
data(totProtAT5)
refLocationProfilesNSA <- locationProfileSetup(profile=protNSA_test,
                               markerList=markerListJadot, numDataCols=9)
round(refLocationProfilesNSA, digits=3)
# Convert NSA reference profiles to Acup to prepare for forming mixtures
refLocationProfilesAcup <- AcupFromNSA(NSA=refLocationProfilesNSA,
                           NstartMaterialFractions=6,
                           totProt=totProtAT5)
round(refLocationProfilesAcup, digits=4)
# Compute mixtures
mixCytoLysoAcup <- proteinMix(AcupRef=refLocationProfilesAcup,
                              increment.prop=0.1,
                              Loc1=1, Loc2=4)
# Convert to relative specific amoounts
mixCytoLysoRSA <- RSAfromAcup(Acup=mixCytoLysoAcup,
                              NstartMaterialFractions=6, totProt=totProtAT5)
# Find RSA transformed reference profiles
refLocationProfilesRSA <- RSAfromNSA(refLocationProfilesNSA,
                          NstartMaterialFractions=6,
                          totProt=totProtAT5)
# Find constrained proportional values
mixCytoLysoCPAfromRSA <- fitCPA(profile=mixCytoLysoRSA,
                                refLocationProfiles=refLocationProfilesRSA,
                                numDataCols=9)
# Plot the mixtures and fitted values and print the error
library(pracma)
mixturePlot(mixProtiProtjCPA=mixCytoLysoCPAfromRSA,
            NstartMaterialFractions=6, Loc1=1, Loc2=4,
            increment.prop=0.1, xaxisLab=TRUE, yaxisLab=TRUE,
            errorReturn = TRUE)
}
