% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixturePlotPanel.r
\name{mixturePlotPanel}
\alias{mixturePlotPanel}
\title{plot mixture of all two compartment profile combinations as panel;
  assumes eight compartments
Also assumes that "refLocationProfilesRSA" has been previously defined}
\usage{
mixturePlotPanel(
  refLocationProfilesAcup,
  totProt,
  NstartMaterialFractions,
  errorReturn = FALSE,
  fitType,
  log2Transf = FALSE,
  eps = 0.001
)
}
\arguments{
\item{refLocationProfilesAcup}{relative amount of a given cellular
compartment protein that ends up in a given centrifugation fraction}

\item{totProt}{vector of amounts starting material in each fraction}

\item{NstartMaterialFractions}{number of starting material fractions}

\item{errorReturn}{return all area-based errors if true}

\item{fitType}{use RSA, NSA, or Acup}

\item{log2Transf}{use log2-transformed values. Default is "FALSE"}

\item{eps}{constant to avoid taking logs of zero}
}
\value{
mixErrorMat a list of errors for all mixtures
}
\description{
plot mixture of all two compartment profile combinations as panel;
  assumes eight compartments
Also assumes that "refLocationProfilesRSA" has been previously defined
}
\examples{
 # See Vignette 4 and Vignette 5 for a full explanation
}
