% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protDistanceCompute.R
\name{nearestProts}
\alias{nearestProts}
\title{Identifies proteins with profiles similar to that of a specified protein.}
\usage{
nearestProts(protName, n.nearest = 5, distProts, protNames, profile)
}
\arguments{
\item{protName}{name of protein to which distances are to be computed}

\item{n.nearest}{number of nearest proteins to list}

\item{distProts}{distance matrix created by, for example}

\item{protNames}{list of all proteins in a dataset}

\item{profile}{dataframe of profiles for proteins}
}
\value{
List of the proteins in data set with profiles closest to 
      that of specified protein (protName)
}
\description{
Uses a distance matrix to identify proteins with profiles nearest 
   to that of a specified protein; can be used to identify proteins 
   with similar subcellular locations.
}
\examples{
data(protNSA_test)
distUse <- dist(protNSA_test[,seq_len(9)], method='euclidean')
protsUse <- names(protNSA_test)
nearestProts(protName='CTSD', n.nearest=10,  distProts=distUse,
  protNames=protsUse, profile=protNSA_test[,seq_len(9)])
}
