% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAtransform.r
\name{NSAfromRSA}
\alias{NSAfromRSA}
\title{convert RSA to NSA}
\usage{
NSAfromRSA(RSA)
}
\arguments{
\item{RSA}{relative specific activity}
}
\value{
NSA A matrix giving the specific amount
       or normalized specific amount
       profiles, either marker profiles from 'cpaSetup' (which requires
       normalized specific amounts as input),
           or a list of many protein profiles
}
\description{
convert RSA to NSA
}
\examples{
data(protNSA_test)
data(totProtAT5)
protAcup <- AcupFromNSA(protNSA_test[,seq_len(9)],
           NstartMaterialFractions = 9, totProt=totProtAT5)
protRSA <- RSAfromAcup(protAcup, NstartMaterialFractions = 9,
          totProt=totProtAT5)
protNSA <- NSAfromRSA(protRSA)
all.equal(protNSA, protNSA_test[,seq_len(9)] )
}
