% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierFind.R
\name{outlierFind_i}
\alias{outlierFind_i}
\title{Find outlier spectra within a protein}
\usage{
outlierFind_i(
  i,
  protClass,
  outlierLevel = "peptide",
  numRefCols,
  numDataCols,
  outlierMeth,
  range,
  proba,
  eps = eps,
  randomError
)
}
\arguments{
\item{i}{unique identifier for protein i}

\item{protClass}{a data frame containing profiles associated 
with either spectra or peptides (see Tutorial 6)}

\item{outlierLevel}{'peptide' for outlier spectra within peptides, or
'protein' for outlier peptides within proteins}

\item{numRefCols}{number of columns (variables) before data columns}

\item{numDataCols}{number of fractions in each profile}

\item{outlierMeth}{boxplot (recommended), scores, or
none (if no outliers are to be reported)}

\item{range}{the range parameter used for identifying outliers
using the boxplot method}

\item{proba}{probability to exclude
outlier for scores method}

\item{eps}{small value to add so that log argument is greater than zero}

\item{randomError}{TRUE if allow it to be random}
}
\value{
New data frame with an additional column of indicators 
          of outlier status
}
\description{
Service function for outlierFind (see help function for outlierFind).
}
\examples{
set.seed(63561)
eps <- 0.029885209
data(spectraNSA_test)
uniqueLabel <- spectraNSA_test$pepId[1]
flagSpectraBox <- outlierFind_i(i=uniqueLabel, protClass=spectraNSA_test,
                      outlierLevel='peptide', numRefCols=5, numDataCols=9,
                      outlierMeth='boxplot', range=3, proba=NA, eps=eps,
                      randomError=TRUE)
str(flagSpectraBox)

}
