% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileSummarize.R
\name{profileSummarize}
\alias{profileSummarize}
\title{Calculates a mean protein or peptide profiles and standard errors}
\usage{
profileSummarize(
  protsCombineCnew,
  numRefCols,
  numDataCols,
  refColsKeep = c(1, 2),
  eps,
  GroupBy = "protId",
  outlierExclude = "spectraAndPeptide",
  cpus = 1,
  singularList = FALSE
)
}
\arguments{
\item{protsCombineCnew}{data frame of profiles (spectra or spectra 
and peptides) with outlier information}

\item{numRefCols}{number of columns preceding the profile data}

\item{numDataCols}{number of fractions in each profile}

\item{refColsKeep}{which reference columns to keep 
(requires columns 1 and 2 for protein and peptide name)}

\item{eps}{small value to add so that log argument is greater than zero}

\item{GroupBy}{‘protId’ if average peptides to give mean protein profile; 
‘peptideId’ if average spectra to give mean peptide profiles}

\item{outlierExclude}{'none', 'spectra', or
'spectraAndpeptide' (default)
 according to exclusion level}

\item{cpus}{1 (default);
if cpus > 1 use BiocParallel with SnowParm
or other multiprocessing method}

\item{singularList}{if TRUE, list fractions associated with a given 
protein or peptide with singular fit from the random effects 
function lmer; default is FALSE}
}
\value{
Mean or weighted mean NSA profiles
}
\description{
ProfileSummarize calculates mean and SE for each channel in each 
    protein or peptide profile using a random effect model (proteins)
    or arithmetic mean (peptides).  A random effect model can avoid 
    dominance of a peptide with a large number spectra.  
    See Tutorial 6 for details.
}
\examples{
set.seed(17356)
eps <- 0.029885209
data(spectraNSA_test)
flagSpectraBox <- outlierFind(protClass=spectraNSA_test,
           outlierLevel='peptide', numRefCols=5, numDataCols=9,
           outlierMeth='boxplot', range=3, eps=eps,
           randomError=TRUE, cpus=1)
pepProfiles <- profileSummarize(protsCombineCnew=flagSpectraBox,
           numRefCols=6, numDataCols=9, refColsKeep=c(1,2,4),eps=eps,
           GroupBy='peptideId', outlierExclude='spectra', cpus=1)
str(pepProfiles, strict.width='cut', width=65)
# Now use multiple processors with specified random number seed
snowParam <- BiocParallel::SnowParam(workers = 2, RNGseed=92883)
BiocParallel::register(snowParam, default=FALSE)
pepProfilesM <- profileSummarize(protsCombineCnew=flagSpectraBox,
           numRefCols=6, numDataCols=9, refColsKeep=c(1,2,4),eps=eps,
           GroupBy='peptideId', outlierExclude='spectra', cpus=2)
str(pepProfilesM, strict.width='cut', width=65)

}
