% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{projSimplex}
\alias{projSimplex}
\title{project to a simplex}
\usage{
projSimplex(y, c = 1)
}
\arguments{
\item{y}{n-dim vector}

\item{c}{constraint; equals one by default}
}
\value{
simplex S_n with sum constrained to 1
}
\description{
internal function; project an n-dim vector y to
the simplex S_n with sum constrained to 1
}
\examples{
y1 <- c(1,1)
p1 <- projSimplex(y1)
p1
y2 <- c(0.1, 0.3)
p2 <- projSimplex(y2)
p2

}
