% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDataPrep.R
\name{proteinDataPrep}
\alias{proteinDataPrep}
\title{Formats a data set of spectral profiles for use with protlocassign}
\usage{
proteinDataPrep(protClass, numRefCols, numDataCols)
}
\arguments{
\item{protClass}{data frame containing protein and peptide names, 
annotation data and profiles}

\item{numRefCols}{number of reference columns
(includes protein and peptide names)}

\item{numDataCols}{number of fractions in each profile}
}
\value{
Data frame properly formatted for protlocassign
}
\description{
The function sorts a data set of peptide-spectrum matches with
   associated profiles and descriptive information, 
   ensuring that spectra are nested within peptides and peptides 
   are nested within proteins so that all peptide and protein data 
   are contiguous. Peptide names (sequences) are prepended with 
   the associated protein name. The function assigns 
   unique numerical identifiers to each protein (protId) and 
   each peptide (pepId). The input data frame must be formatted 
   so that the first column contains protein names and the second 
   peptide names (typically sequences).  
   This is followed by a series of columns containing descriptive 
   information and profile data.
}
\examples{
data(QuantPSM_test)
specNSA_out <- proteinDataPrep(QuantPSM_test, numRefCols=5,numDataCols=9)
head(specNSA_out)
}
