% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinMix.r
\name{proteinMix}
\alias{proteinMix}
\title{Simulate mixture protein resident in two compartments}
\usage{
proteinMix(AcupRef, Loc1, Loc2, increment.prop = 0.1)
}
\arguments{
\item{AcupRef}{amount of given protein in fraction /
amount of given protein in starting material}

\item{Loc1}{row number of one compartment}

\item{Loc2}{row number of other compartment}

\item{increment.prop}{fraction increment from 0 to 1}
}
\value{
mixAmount relative amounts of proteins in the fractions
}
\description{
Compute the profile of proteins distributed between two compartments in
 specified proportions
}
\examples{
data(protNSA_test)
data(markerListJadot)
data(totProtAT5)
refLocationProfilesNSA <- locationProfileSetup(profile=protNSA_test,
                                markerList=markerListJadot, numDataCols=9)
round(refLocationProfilesNSA, digits=3)
# Convert NSA reference profiles to Acup to prepare for forming mixtures
refLocationProfilesAcup <-
     AcupFromNSA(NSA=refLocationProfilesNSA, NstartMaterialFractions=6,
     totProt=totProtAT5)
round(refLocationProfilesAcup, digits=4)
# Compute mixtures
mixCytoLysoAcup <- proteinMix(AcupRef=refLocationProfilesAcup,
                              increment.prop=0.1,
                              Loc1=1, Loc2=4)
}
