% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{protIndex}
\alias{protIndex}
\title{Index of protein name}
\usage{
protIndex(protName, profile, exactMatch = FALSE)
}
\arguments{
\item{protName}{name of protein to search for}

\item{profile}{data frame of specified protein(row name) profiles}

\item{exactMatch}{default is FALSE}
}
\value{
The protein name and its index (row in profile)
}
\description{
Return index of a protein name and (if exactMatch=FALSE, the default) 
indices of proteins starting with characters that match the string 
given in 'protName'. If exactMatch=TRUE, return only the index of 
a protein name that exactly matches 'protName'.
}
\examples{
data(protNSA_test)
protIndex('TLN1', profile=protNSA_test)
protIndex('TLN', profile=protNSA_test)
}
