% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protNSA_test.r
\docType{data}
\name{protNSA_test}
\alias{protNSA_test}
\title{Spectral-level test data set}
\format{
a test data set

\describe{
  \item{\code{N}}{a numeric vector}
  \item{\code{M}}{a numeric vector}
  \item{\code{L1}}{a numeric vector}
 \item{\code{L2}}{a numeric vector}
  \item{\code{P}}{a numeric vector}
  \item{\code{S}}{a numeric vector}
  \item{\code{Nyc1}}{a numeric vector}
  \item{\code{Nyc2}}{a numeric vector}
  \item{\code{Nyc3}}{a numeric vector}
  \item{\code{Nspectra}}{a numeric vector}
  \item{\code{Npep}}{a numeric vector}
}
}
\usage{
data(protNSA_test)
}
\description{
Test data set with  TLN1, TLN2, AIF1, NSA mean profiles
  as well as mean profiles of all of the proteins 
  listed in "markerListJadot".
  Data from the TMT MS2 data, AT5 experiment; see Tannous et al
}
\examples{
data(protNSA_test)
str(protNSA_test)
}
\references{
Tannous, A.; Boonen, M.; Zheng, H.; Zhao, C.;
 Germain, C. J.; Moore, D. F.;
 Sleat, D. E.; Jadot, M.; Lobel, P., Comparative Analysis of
 Quantitative Mass Spectrometric Methods for Subcellular Proteomics.
 J Proteome Res 2020, 19, (4), 1718-1730.
}
\keyword{datasets}
