% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protPepProfilePlot.R
\name{protPepPlotfun}
\alias{protPepPlotfun}
\title{Plot reference compartment profiles overlaid with the profile 
  of a single protein and its component peptides}
\usage{
protPepPlotfun(
  protName,
  protProfile,
  Nspectra = TRUE,
  pepProfile = NULL,
  numRefCols,
  numDataCols,
  n.compartments = 8,
  refLocationProfiles,
  assignPropsMat,
  propCI = FALSE,
  transType = "",
  yAxisLabel = ""
)
}
\arguments{
\item{protName}{name of the protein to plot}

\item{protProfile}{data frame containing protein profiles}

\item{Nspectra}{number of spectra associated to a given peptide}

\item{pepProfile}{data frame containing peptide profiles}

\item{numRefCols}{number of reference columns
(preceding the data profile columns)}

\item{numDataCols}{number of fractions in each profile}

\item{n.compartments}{number of compartments}

\item{refLocationProfiles}{data frame of profiles for the  reference compartments}

\item{assignPropsMat}{data frame containing CPA values}

\item{propCI}{TRUE if lower and upper confidence intervals are included 
in assignPropsMat (not currently implemented)}

\item{transType}{label for y-axis on each individual plot (default none))}

\item{yAxisLabel}{label for y-axis of entire panel}
}
\value{
Plot of protein, peptide, and reference profiles
}
\description{
This function creates a panel of plots, one for each compartment.  
     Each plot contains a reference compartment profile overlaid 
     with the profiles of one protein of interest and its component
      peptides as well as the CPA value for that compartment.
}
\examples{
  # See Tutorial 6 for a full explanation
}
