% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protProfilePlot.R
\name{protPlotfun}
\alias{protPlotfun}
\title{Plot reference compartment profiles overlaid with the profile of 
  a single protein}
\usage{
protPlotfun(
  protName,
  profile,
  numDataCols = 9,
  n.compartments = 8,
  refLocationProfiles,
  assignPropsMat,
  transType = "",
  yAxisLabel = ""
)
}
\arguments{
\item{protName}{name of the protein to plot}

\item{profile}{data frame of specified protein(row name) profiles}

\item{numDataCols}{number of fractions in each profile}

\item{n.compartments}{number of compartments (8 in Jadot data)}

\item{refLocationProfiles}{data frame of profiles for the 
reference compartments}

\item{assignPropsMat}{data frame containing CPA estimates}

\item{transType}{label for y-axis on each individual plot (default none))}

\item{yAxisLabel}{label for y-axis of entire panel}
}
\value{
A panel of profile plots
}
\description{
This function creates a panel of plots, one for each compartment.  
   Each plot contains a reference compartment profile overlaid with 
   the profile of one protein of interest and its CPA 
   value for that compartment.
}
\examples{
data(protNSA_test)
data(markerListJadot)
data(refLocProfNSA)
protCPAfromNSA_out <- fitCPA(profile=protNSA_test,
                              refLocationProfiles=refLocProfNSA,
                              numDataCols=9)
protPlotfun(protName="TLN1", profile=protNSA_test,
            numDataCols=9, n.compartments=8,
            refLocationProfiles=refLocProfNSA,
            assignPropsMat=protCPAfromNSA_out,
            yAxisLabel="Normalized Specific Amount")
}
