% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantPSM_test.r
\docType{data}
\name{QuantPSM_test}
\alias{QuantPSM_test}
\title{Annotated peptide/spectra/protein test data set}
\format{
a test data set of NSA spectra

\describe{
  \item{\code{Protein.Gene}}{protein name}
  \item{\code{PeptideSequence}}{peptide}
  \item{\code{SpectrumId}}{identification label of spectrum}
  \item{\code{PeptidesStartPostionInProtein}}{number of start position}
  \item{\code{modifications}}{peptide modification list}
  \item{\code{N}}{a numeric vector}
  \item{\code{M}}{a numeric vector}
  \item{\code{L1}}{a numeric vector}
  \item{\code{L2}}{a numeric vector}
  \item{\code{P}}{a numeric vector}
  \item{\code{S}}{a numeric vector}
  \item{\code{Nyc1}}{a numeric vector}
  \item{\code{Nyc2}}{a numeric vector}
  \item{\code{Nyc3}}{a numeric vector}
}
}
\usage{
data(QuantPSM_test)
}
\description{
Test data set of annotated peptide spectrum matches and associated 
NSA profiles, used to construct spectraNSA_test data set.
Drawn from data from the TMT MS2 data, AT5 experiment; see Tannous et al
}
\examples{
data(QuantPSM_test)
str(QuantPSM_test)
}
\references{
Tannous, A.; Boonen, M.; Zheng, H.; Zhao, C.;
 Germain, C. J.; Moore, D. F.;
 Sleat, D. E.; Jadot, M.; Lobel, P., Comparative Analysis of
 Quantitative Mass Spectrometric Methods for Subcellular Proteomics.
 J Proteome Res 2020, 19, (4), 1718-1730.
}
\keyword{datasets}
