% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refLocProfNSA.R
\docType{data}
\name{refLocProfNSA}
\alias{refLocProfNSA}
\title{NSA reference profiles}
\format{
List of profiles
}
\usage{
data(refLocProfNSA)
}
\value{
profiles for subcellular compartments based on NSA means

\describe{
  \item{\code{N}}{a numeric vector}
  \item{\code{M}}{a numeric vector}
  \item{\code{L1}}{a numeric vector}
 \item{\code{L2}}{a numeric vector}
  \item{\code{P}}{a numeric vector}
  \item{\code{S}}{a numeric vector}
  \item{\code{Nyc1}}{a numeric vector}
  \item{\code{Nyc2}}{a numeric vector}
  \item{\code{Nyc3}}{a numeric vector}
}
}
\description{
Profiles for the subcellular compartments based on NSA profile means of the 
   individual reference proteins that represent each compartment. 
   This function produces a matrix that has one row for each 
   compartment and one column for each fraction that comprises the profile.
  Data from the TMT MS2 data, AT5 experiment; see Tannous et al.
  See help file for `locationProfileSetup` for more information.
}
\examples{
data(refLocProfNSA)
round(refLocProfNSA, digits=4)
}
\references{
Tannous, A.; Boonen, M.; Zheng, H.; Zhao, C.;
 Germain, C. J.; Moore, D. F.;
 Sleat, D. E.; Jadot, M.; Lobel, P., Comparative Analysis of
 Quantitative Mass Spectrometric Methods for Subcellular Proteomics.
 J Proteome Res 2020, 19, (4), 1718-1730.
}
\keyword{datasets}
