\name{totProtAT5}
\alias{totProtAT5}
\docType{data}
\title{Amount of total protein in each fraction derived from
a gram of tissue homogenate in experiment AT5 from Tannous et al. 2000
}
\description{
This is a vector with nine values in the same order as protNSA_test: 
there is a value for each of the six differential centrifugation 
fractions (N, M, L1, L2, P and S) and the three Nycodenz channels. 
Note that the sum of the six differential centrifugation fractions 
is used to estimate the total protein in the tissue homogenate.  
}
\usage{data("totProtAT5")}
\format{
  The format is:
 names: "N" "M" "L1" "L2" "P" "S" "Nyc.1" "Nyc.2" "Nyc.3""
}

\references{
A Tannous et al. (2020)
}
\examples{
data(totProtAT5)
totProtAT5
}
\keyword{datasets}
