% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{vecUnitize}
\alias{vecUnitize}
\title{unitize vector}
\usage{
vecUnitize(xx)
}
\arguments{
\item{xx}{vector}
}
\value{
normalized vector of unit length
}
\description{
normalize a vector to have unit length
}
\examples{
xx <- c(0.5, 0.1, 0.6, 0.9)
vecUnitize(xx)
}
