% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAtransform.r
\name{AcupFromNSA}
\alias{AcupFromNSA}
\title{Convert NSA to Acup}
\usage{
AcupFromNSA(NSA, NstartMaterialFractions, totProt = NULL)
}
\arguments{
\item{NSA}{A matrix giving normalized specific amount profiles 
(see help file for protNSA_test)}

\item{NstartMaterialFractions}{Number of fractions that reconstitute 
the starting material, e.g., a complete set of differential 
centrifugation fractions.  For experiment AT5, it is 6 
( N, M, L1, L2, P, and S).}

\item{totProt}{Total protein amount (derived from a given amount 
of starting material) in each of the fractions comprising the profile}
}
\value{
Acup profiles;
  amount of given protein in fraction /
  amount of that given protein in starting material
  (see help file for protAcup_test).
}
\description{
Convert NSA (normalized specific amount) profiles to Acup 
(relative amount) profiles
}
\examples{
data(protNSA_test)
data(totProtAT5)
protAcup_out1 <- AcupFromNSA(protNSA_test[,seq_len(9)],
   NstartMaterialFractions = 6, totProt=totProtAT5)
head(protAcup_out1)
}
