% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAtransform.R
\name{AcupFromRSA}
\alias{AcupFromRSA}
\title{Convert RSA to Acup}
\usage{
AcupFromRSA(RSA, NstartMaterialFractions, totProt = NULL)
}
\arguments{
\item{RSA}{matrix of relative specific amount profiles 
(see help file for protRSA_test)}

\item{NstartMaterialFractions}{number of fractions that reconstitute 
the starting material, e.g., a complete set of differential 
centrifugation fractions.  For experiment AT5, it is 6 
( N, M, L1, L2, P, and S).}

\item{totProt}{vector of total protein amounts (derived from 
a given amount of starting material) in each of the 
fractions comprising the profile}
}
\value{
Acup profiles: amount of given protein in fraction / 
        amount of that given protein in starting material 
        (see help file protAcup_test).
}
\description{
Convert RSA (relative specific amount) profiles to Acup
(relative amount) profiles
}
\examples{
data(protRSA_test)
data(totProtAT5)
protAcup_out2 <- AcupFromRSA(protRSA_test[,seq_len(9)],
            NstartMaterialFractions = 6, totProt=totProtAT5)
round(head(protAcup_out2), digits=4)
str(protAcup_out2)
}
