% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{assignCPAloc}
\alias{assignCPAloc}
\title{Assign to single subcellular location}
\usage{
assignCPAloc(assignLocProps, cutoff = 0.8, Locations)
}
\arguments{
\item{assignLocProps}{matrix of CPA estimates for each protein}

\item{cutoff}{cutoff for assigning a protein to a location}

\item{Locations}{list of subcellular locations}
}
\value{
Single compartment where most of protein resides, 
       undefined if largest CPA value is <0.5
}
\description{
Assign proteins to the most prevalent subcellular location 
(meeting a specific cutoff) using CPA estimates,
with minimum value being 0.5.
}
\examples{
data(protNSA_test)
data(markerListJadot)
data(refLocProfNSA)
protCPAfromNSA_test <- fitCPA(profile=protNSA_test,
         refLocationProfiles=refLocProfNSA,
         numDataCols=9)
Locations <- unique(markerListJadot$referenceCompartment)

# write table listing predominant CPA assignment of each protein in data set
table(apply(protCPAfromNSA_test[,1:8],1,assignCPAloc,
         cutoff=0.8, Locations=Locations))
}
