% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{fCPAone}
\alias{fCPAone}
\title{CPA for protein i}
\usage{
fCPAone(
  profile,
  refLocationProfiles,
  numDataCols,
  startProps = NULL,
  maxit = 10000,
  ind.vary = NULL,
  minVal = FALSE
)
}
\arguments{
\item{profile}{data frame of one protein name (row name) and
relative abundance levels for that protein}

\item{refLocationProfiles}{A matrix giving the abundance
level profiles of the subcellular locations}

\item{numDataCols}{Number of channels of abundance levels}

\item{startProps}{starting values for proportional assignments;
set equal if this is null (default)}

\item{maxit}{maximum number of iterations (default is 10000)}

\item{ind.vary}{if not NULL, indexes of proportions allowed to vary with 
others constrained to zero}

\item{minVal}{default is false. If true, return minimum value of
goodness of fit}
}
\value{
assignProbsOut  Vector of proportional assignments of
   each protein to compartments. Also returns variables
   nspectra and npeptides if they are included in the profile input.
   If ind.vary is specified, only the 
   referenced proportion CPA estimates are returned.
}
\description{
Carry out constrained proportional assignment for protein i;
   service function for fitCPA
}
\examples{
data(protRSA_test)
data(refLocProfRSA_test)

protCPAfromRSA_i_out1 <- fCPAone(profile=protRSA_test[1,],
                          refLocationProfiles=refLocProfRSA,
                          numDataCols=9, startProps=NULL,
                          maxit=10000,
                         ind.vary=NULL, minVal=FALSE)
round(protCPAfromRSA_i_out1, digits=4)
}
