% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{fCPAone}
\alias{fCPAone}
\title{CPA for protein i}
\usage{
fCPAone(
  profile,
  refLocationProfiles,
  numDataCols,
  startProps = NULL,
  maxit = 10000,
  ind.vary = NULL,
  minVal = FALSE
)
}
\arguments{
\item{profile}{data frame of one protein name (row name) and
relative abundance levels for that protein}

\item{refLocationProfiles}{A matrix giving the abundance
level profiles of the subcellular locations}

\item{numDataCols}{Number of channels of abundance levels}

\item{startProps}{starting values for proportional assignments;
set equal if this is null (default)}

\item{maxit}{maximum number of iterations (default is 10000)}

\item{ind.vary}{if not NULL, indexes of proportions allowed to vary with 
others constrained to zero}

\item{minVal}{default is false. If true, return minimum value of
goodness of fit}
}
\value{
assignProbsOut  Data frame of proportionate assignments of
   each protein to compartments. also nspectra and npeptides if in
   profile input
   If ind.vary specified, only the referenced proportion cpa
   estimates are returned
}
\description{
Carry out constrained proportional assignment for protein i;
   service function for fitCPA
}
\examples{
data(protNSA_test)
data(markerListJadot)
nTestProts <- nrow(protNSA_test)
refLocationProfilesNSA <- locationProfileSetup(profile=protNSA_test,
    markerList=markerListJadot, numDataCols=9)
protCPAfromNSA_test <- fCPAone(profile=protNSA_test[1,],
                          refLocationProfiles=refLocationProfilesNSA,
                          numDataCols=9, startProps=NULL,
                          maxit=10000,
                         ind.vary=NULL, minVal=FALSE)
}
