% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCPAsubsets.r
\name{fCPAsubsets}
\alias{fCPAsubsets}
\title{constrained CPA}
\usage{
fCPAsubsets(
  profile,
  refLocationProfiles,
  numDataCols,
  startProps = NULL,
  showProgress = TRUE,
  maxit = 10000,
  nCPAcomparts = 2
)
}
\arguments{
\item{profile}{a vector containing a specified protein (row name) 
profile ‘Nspectra’ and ‘Npep’, if present, will be removed}

\item{refLocationProfiles}{data frame of profiles for the
reference compartments}

\item{numDataCols}{number of fractions in each profile}

\item{startProps}{starting values for proportional assignments;
set equal if this is null (default)}

\item{showProgress}{default is TRUE}

\item{maxit}{maximum number of iterations (default is 10000)}

\item{nCPAcomparts}{number of compartments to fit restricted CPA;
remaining proportions are fixed at zero}
}
\value{
Data frame with CPA estimates for every combination
       ordered by value, with best fit listed first
}
\description{
Carry out constrained proportional assignment for protein
 number i in profile
 with constraint being the number of compartments that are allowed to
 vary freely; other compartment proportions are fixed at zero
}
\examples{
data(protRSA_test)
data(refLocProfRSA)
protCPAfromRSA_out2 <- fCPAsubsets(profile=protRSA_test[1,],
                         refLocationProfiles=refLocProfRSA,
                         numDataCols=9, startProps=NULL, nCPAcomparts=2)
round(head(protCPAfromRSA_out2), digits=4)
}
