% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{fitCPA}
\alias{fitCPA}
\title{CPA for a set of proteins}
\usage{
fitCPA(
  profile,
  refLocationProfiles,
  numDataCols,
  startProps = NULL,
  maxit = 10000,
  showProgress = TRUE,
  ind.vary = NULL,
  minVal = FALSE
)
}
\arguments{
\item{profile}{a data frame of protein names (row names) and
relative abundance levels.}

\item{refLocationProfiles}{A matrix giving the abundance level
profiles of the subcellular locations}

\item{numDataCols}{Number of channels of abundance levels}

\item{startProps}{starting values for proportional assignments;
set equal if this is null (default)}

\item{maxit}{maximum number of iterations (default is 10000)}

\item{showProgress}{print out progress if TRUE, the default}

\item{ind.vary}{if not NULL, indexes of parameters to allow to vary; 
remaining parameters are fixed at zero}

\item{minVal}{default is false. If true, return minimum value of
goodness of fit}
}
\value{
assignProbsOut  Data frame of proportional assignments
      of each protein to compartments
}
\description{
Carry out constrained proportional assignment on a profiles
 of set of proteins
}
\examples{
data(protRSA_test)
data(refLocProfRSA)
protCPAfromRSA_out <- fitCPA(profile=protRSA_test,
                           refLocationProfiles=refLocProfRSA,
                           numDataCols=9)
round(head(protCPAfromRSA_out), digits=4)                           
}
