% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{fitCPA}
\alias{fitCPA}
\title{CPA for a set of proteins}
\usage{
fitCPA(
  profile,
  refLocationProfiles,
  numDataCols,
  startProps = NULL,
  maxit = 10000,
  showProgress = TRUE,
  ind.vary = NULL,
  minVal = FALSE
)
}
\arguments{
\item{profile}{data frame of specified protein(row name) profiles}

\item{refLocationProfiles}{data frame of profiles for 
the reference compartments}

\item{numDataCols}{number of fractions in each profile}

\item{startProps}{starting values for proportional assignments;
set equal if this is null (default)}

\item{maxit}{maximum number of iterations (default is 10000)}

\item{showProgress}{print out progress if TRUE, the default}

\item{ind.vary}{if not NULL, indexes of parameters to allow to vary; 
remaining parameters are fixed at zero}

\item{minVal}{default is FALSE. If TRUE, return minimum value of
goodness of fit}
}
\value{
Data frame of CPA estimates for each protein
}
\description{
Carry out constrained proportional assignment on profiles
 of a set of proteins
}
\examples{
data(protRSA_test)
data(refLocProfRSA)
protCPAfromRSA_out <- fitCPA(profile=protRSA_test,
                           refLocationProfiles=refLocProfRSA,
                           numDataCols=9)
# Note that the profile of one protein, AIF1, contains missing values
# which causes the cpa routine to generate a nonconvergence message                           
round(head(protCPAfromRSA_out), digits=4)                           
}
