% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{locationProfileSetup}
\alias{locationProfileSetup}
\title{Set up reference profiles for constrained proportional assignment}
\usage{
locationProfileSetup(profile, markerList, numDataCols)
}
\arguments{
\item{profile}{data frame of protein identifiers (protName)
and their relative abundance in centrifugation fractions.}

\item{markerList}{List of reference proteins and their subcellular locations}

\item{numDataCols}{Number of channels of abundance levels}
}
\value{
A matrix refLocationProfiles giving the abundance level profiles of
        the subcellular locations
}
\description{
Set up reference profiles for constrained proportional assignment
}
\examples{
#data(protNSA_AT5tmtMS2)
data(protNSA_test)
data(markerListJadot)
refLocationProfilesNSA <- locationProfileSetup(profile=protNSA_test,
  markerList=markerListJadot, numDataCols=9)
}
