% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerProfilePlot.r
\name{markerProfilePlot}
\alias{markerProfilePlot}
\title{Plot profiles of reference proteins}
\usage{
markerProfilePlot(
  refLoc,
  profile,
  markerList,
  refLocationProfiles,
  ylab = "",
  refProtPlot = NULL
)
}
\arguments{
\item{refLoc}{the name of the reference subcellular
compartment to plot}

\item{profile}{data frame of protein names (as row names)
and relative abundance levels.}

\item{markerList}{List of reference proteins}

\item{refLocationProfiles}{A matrix refLocationProfiles giving
the abundance level profiles of the subcellular locations}

\item{ylab}{label for y=axis, eg, 'NSA'}

\item{refProtPlot}{indices of reference proteins to plot; default is NULL}
}
\value{
plot of profiles of reference (marker) proteins
}
\description{
This function plots profiles of reference proteins
   and also the average profile for each compartment
}
\examples{
data(protNSA_AT5tmtMS2)
data(markerListJadot)
refLocationProfilesNSA <- locationProfileSetup(profile=protNSA_AT5tmtMS2,
   markerList=markerListJadot, numDataCols=9)
markerProfilePlot(refLoc='PM', profile=protNSA_AT5tmtMS2,
   markerList=markerListJadot,
   refLocationProfiles=refLocationProfilesNSA, ylab='NSA')
}
