% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileSummarize.r
\name{meansByProteins}
\alias{meansByProteins}
\title{Find mean profile for a protein or peptide}
\usage{
meansByProteins(
  i,
  uniqueLabel,
  protsCombineCnew,
  numRefCols,
  numDataCols,
  GroupBy,
  eps,
  outlierExclude,
  singularList = FALSE
)
}
\arguments{
\item{i}{protID or pepID (only one)}

\item{uniqueLabel}{vector of either
protId or pepId as specified by 'GroupBy' in
profileSummarize}

\item{protsCombineCnew}{dataframe of protein profiles}

\item{numRefCols}{number of columns preceding the data}

\item{numDataCols}{number of fractions in each profile}

\item{GroupBy}{'protId' if average by
protein; 'peptideId' if average by peptide}

\item{eps}{small value to add so that log
argument is greater than zero}

\item{outlierExclude}{none, spectra, or
spectraAndPeptide}

\item{singularList}{If TRUE, list proteins and channels with singular
lmer fit results; default is FALSE}
}
\value{
Coefficient
     estimates, numbers of peptides and spectra, and
     id for protein i

estimated means for all channels, their standard errors,
and peptide or protein info
}
\description{
Service function for profileSummarize. Find mean profiles using a 
  random effects model for nested data.
}
\examples{
set.seed(17356)
eps <- 0.029885209
data(spectraNSA_test)
flagSpectraBox <- outlierFind(protClass=spectraNSA_test,
                              outlierLevel='peptide', numRefCols=5,
                              numDataCols=9,
                              outlierMeth='boxplot', range=3, eps=eps,
                              randomError=TRUE)
                              
uniqueLabel <- flagSpectraBox$pepId
pepProfile_i_out <- meansByProteins(i=uniqueLabel[1], 
          uniqueLabel=uniqueLabel,
          protsCombineCnew=flagSpectraBox,
          numRefCols=6, numDataCols=9, GroupBy="peptideId",eps=eps,
          outlierExclude='spectra')
round(pepProfile_i_out, digits=3)
}
