% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureAreaError.r
\name{mixtureAreaError}
\alias{mixtureAreaError}
\title{Area-based error of mixtures}
\usage{
mixtureAreaError(
  mixProtiProtjCPA,
  NstartMaterialFractions = 6,
  Loc1,
  Loc2,
  increment.prop = 0.1
)
}
\arguments{
\item{mixProtiProtjCPA}{data frame of CPA estimates for each set 
of two compartment mixtures}

\item{NstartMaterialFractions}{Number of fractions that reconstitute 
the starting material, e.g., a complete set of differential 
centrifugation fractions.  For experiment AT5, it is 6 
( N, M, L1, L2, P, and S).}

\item{Loc1}{row number of subcellular location 1 of mixture}

\item{Loc2}{row number of subcellular location 2 of mixture}

\item{increment.prop}{increments in proportion residing in Loc1
(from 0 to 1); Default is 0.1}
}
\value{
The area between predicted and observed curves
}
\description{
Compute area-based error for accuracy of CPA estimates for 
   sets of simulated proteins distributed between two compartments.
}
\examples{
data(refLocProfAcup)
data(refLocProfNSA)
data(totProtAT5)

# Compute relative amount of each theoretical protein that resides 
#    in each fraction in a mixture set
mixCytoLysoAcup <- proteinMix(AcupRef=refLocProfAcup,
                              increment.prop=0.1,
                              Loc1=1, Loc2=4)
                              
# Convert theoretical protein profiles to NSA
mixCytoLysoNSA <- NSAfromAcup(Acup=mixCytoLysoAcup,
                           NstartMaterialFractions=6, totProt=totProtAT5)
                              
# Find constrained proportional values
mixCytoLysoCPAfromNSA <- fitCPA(profile=mixCytoLysoNSA,
                                refLocationProfiles=refLocProfNSA,
                                numDataCols=9)
# calculate the mixture error
mixtureAreaError(mixProtiProtjCPA=mixCytoLysoCPAfromNSA,
            NstartMaterialFractions=6, Loc1=1, Loc2=4,
            increment.prop=0.1)
}
