% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureAreaError.r
\name{mixtureAreaError}
\alias{mixtureAreaError}
\title{Area-based error of mixtures}
\usage{
mixtureAreaError(
  mixProtiProtjCPA,
  NstartMaterialFractions = 6,
  Loc1,
  Loc2,
  increment.prop = 0.1
)
}
\arguments{
\item{mixProtiProtjCPA}{data frame of CPA estimated proportions
for each mixture}

\item{NstartMaterialFractions}{Number of fractions that reconstitute 
the starting material, e.g., a complete set of differential 
centrifugation fractions.  For experiment AT5, it is 6 
( N, M, L1, L2, P, and S).}

\item{Loc1}{row number of subcellular location 1 of mixture}

\item{Loc2}{row number of subcellular location 2 of mixture}

\item{increment.prop}{increment for computation; default is 0.1}
}
\value{
error, the area between predicted and observed curves
}
\description{
Compute area-based error for mixture of two
   compartment profiles
}
\examples{
data(protNSA_test)
data(markerListJadot)
data(totProtAT5)
refLocationProfilesNSA <- locationProfileSetup(profile=protNSA_test,
                              markerList=markerListJadot, numDataCols=9)
round(refLocationProfilesNSA, digits=3)
# Convert NSA reference profiles to Acup to prepare for forming mixtures
refLocationProfilesAcup <- AcupFromNSA(NSA=refLocationProfilesNSA,
               NstartMaterialFractions=6,
               totProt=totProtAT5)
round(refLocationProfilesAcup, digits=4)
# Compute mixtures
mixCytoLysoAcup <- proteinMix(AcupRef=refLocationProfilesAcup,
                              increment.prop=0.1,
                              Loc1=1, Loc2=4)
# Convert to relative specific amoounts
mixCytoLysoRSA <- RSAfromAcup(Acup=mixCytoLysoAcup,
                              NstartMaterialFractions=6, totProt=totProtAT5)
# Find RSA transformed reference profiles
refLocationProfilesRSA <- RSAfromNSA(refLocationProfilesNSA,
                          NstartMaterialFractions=6,
                          totProt=totProtAT5)
# Find constrained proportional values
mixCytoLysoCPAfromRSA <- fitCPA(profile=mixCytoLysoRSA,
                                refLocationProfiles=refLocationProfilesRSA,
                                numDataCols=9)
# calculate the mixture error
mixtureAreaError(mixProtiProtjCPA=mixCytoLysoCPAfromRSA,
            NstartMaterialFractions=6, Loc1=1, Loc2=4,
            increment.prop=0.1)
}
