% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureHeatMap.r
\name{mixtureHeatMap}
\alias{mixtureHeatMap}
\title{Heat map of mixture CPA errors}
\usage{
mixtureHeatMap(Acup, totProt, NstartMaterialFractions = 6, eps = 0.001)
}
\arguments{
\item{Acup}{relative amount of a given cellular
compartment protein that ends up in a given centrifugation fraction}

\item{totProt}{vector of amounts starting material in each fraction}

\item{NstartMaterialFractions}{Number of fractions that reconstitute 
the starting material, e.g., a complete set of differential 
centrifugation fractions.  For experiment AT5, it is 6 
( N, M, L1, L2, P, and S).}

\item{eps}{small positive constant to add before taking a log transformation}
}
\value{
color heat map of errors
}
\description{
Produce heat map of errors of mixture of all two compartment profile 
combinations as panel; assumes eight compartments
}
\examples{
  # See See Vignette 4 for a full explanation
}
