% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureHeatMap.R
\name{mixtureHeatMap}
\alias{mixtureHeatMap}
\title{Heat map of mixture CPA errors}
\usage{
mixtureHeatMap(
  Acup,
  totProt,
  NstartMaterialFractions = 6,
  eps = 0.001,
  errorListIn = NULL,
  errorListReturn = FALSE
)
}
\arguments{
\item{Acup}{Acup profiles of reference compartments}

\item{totProt}{vector of total protein amounts (derived from a given 
amount of starting material) in each of the fractions 
comprising the profile}

\item{NstartMaterialFractions}{Number of fractions that reconstitute 
the starting material, e.g., a complete set of differential 
centrifugation fractions.  For experiment AT5, it is 6 
( N, M, L1, L2, P, and S).}

\item{eps}{small value to add so that log argument is greater than zero}

\item{errorListIn}{If not NULL, pre-computed errors are passed to the
function to save computing time for display purposes.
 The format must be as
in the $errorList component in the output. Default is NULL}

\item{errorListReturn}{If TRUE, return data frame of mixture errors
if FALSE (the default) return matrix of total errors}
}
\value{
Plot of heat map of errors of CPA estimates for pairs of 
      simulated protein mixtures represented by different types of 
      profiles and a table of total errors for each type of transformation.
        Errors are displayed in a 3x2 output where the top row of the 
        left, middle and right columns represent RSA, NSA, 
        and Acup profiles, respectively and the bottom row represents 
        a log2-transformation of these profiles. 
        By default return a matrix total errors. If errorListShow=TRUE,
        return data frame of mixture errors
}
\description{
Produce heat map of errors in CPA estimates for fitting simulated 
   proteins distributed between all pairs of compartments using 
   different types of profiles (RSA, NSA, and Acup; 
   linear and log2-transformed). 
   Formatted for use with up to eight compartments , 
   which produces 28 (8 chose 2) pairs of combinations of compartments. 
   Also writes a table of total assignment errors for each type of 
   transformation.
}
\examples{
\donttest{
data(totProtAT5)
data(refLocProfAcup)
par(mfrow=c(1,1))
errorResult <- mixtureHeatMap(Acup=refLocProfAcup, totProt=totProtAT5)
round(head(errorResult), digits=3)
}
# Note that the profile of one protein, AIF1, contains missing values
# which causes the cpa routine to generate a nonconvergence message 
}
