% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixturePlot.r
\name{mixturePlot}
\alias{mixturePlot}
\title{Plot set of simulated proteins distributed between two compartments.}
\usage{
mixturePlot(
  mixProtiProtjCPA,
  NstartMaterialFractions = 6,
  Loc1,
  Loc2,
  increment.prop = 0.1,
  errorReturn = FALSE,
  subTitle = NULL,
  xaxisLab = TRUE,
  yaxisLab = TRUE
)
}
\arguments{
\item{mixProtiProtjCPA}{data frame of CPA estimates for each 
set of two compartment mixtures}

\item{NstartMaterialFractions}{Number of fractions that reconstitute 
the starting material, e.g., a complete set of differential 
centrifugation fractions.  For experiment AT5, it is 6 
( N, M, L1, L2, P, and S).}

\item{Loc1}{row number of subcellular location 1 of mixture}

\item{Loc2}{row number of subcellular location 2 of mixture}

\item{increment.prop}{increments in proportion residing in 
(from 0 to 1);
Default is 0.1}

\item{errorReturn}{return area of error region if TRUE}

\item{subTitle}{subtitle for plot if present; NULL if not (default)}

\item{xaxisLab}{plot label for x-axis if TRUE}

\item{yaxisLab}{plot label for y-axis if TRUE}
}
\value{
Plot CPA estimates versus true mixture proportions.
}
\description{
For a set of simulated proteins distributed between two compartments, 
      plot the CPA estimates from a given type of profile
      versus true mixture proportions.
}
\examples{
data(protNSA_test)
data(markerListJadot)
data(totProtAT5)
data(refLocProfAcup)
data(refLocProfNSA)

mixCytoLysoAcup <- proteinMix(AcupRef= refLocProfAcup,
                              increment.prop=0.1,
                              Loc1=1, Loc2=4)
# Convert mixture Acup profiles to NSA profiles
mixCytoLysoNSA <- NSAfromAcup(Acup=mixCytoLysoAcup,
                      NstartMaterialFractions=6, totProt=totProtAT5)
                      
# Find constrained proportional values
mixCytoLysoCPAfromNSA <- fitCPA(profile=mixCytoLysoNSA,
                                refLocationProfiles=refLocProfNSA,
                                numDataCols=9)

# Plot the mixtures and fitted values and print the error
library(pracma)
mixturePlot(mixProtiProtjCPA=mixCytoLysoCPAfromNSA,
            NstartMaterialFractions=6, Loc1=1, Loc2=4,
            increment.prop=0.1, xaxisLab=TRUE, yaxisLab=TRUE,
            errorReturn = TRUE)
            
}
