% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixturePlotPanel.r
\name{mixturePlotPanel}
\alias{mixturePlotPanel}
\title{Mixture plots of all pairwise simulated combinations}
\usage{
mixturePlotPanel(
  refLocationProfilesAcup,
  totProt,
  NstartMaterialFractions,
  errorReturn = FALSE,
  fitType,
  log2Transf = FALSE,
  eps = 0.001
)
}
\arguments{
\item{refLocationProfilesAcup}{relative amount of a given cellular
compartment protein that ends up in a given centrifugation fraction}

\item{totProt}{vector of amounts starting material in each fraction}

\item{NstartMaterialFractions}{Number of fractions that reconstitute 
the starting material, e.g., a complete set of differential 
centrifugation fractions.  For experiment AT5, it is 6 
( N, M, L1, L2, P, and S).}

\item{errorReturn}{return all area-based errors if true}

\item{fitType}{use RSA, NSA, or Acup}

\item{log2Transf}{use log2-transformed values. Default is "FALSE"}

\item{eps}{constant to avoid taking logs of zero}
}
\value{
mixErrorMat a list of errors for all mixtures
}
\description{
Plot mixture of all two compartment profile combinations as panel;
  Assumes eight compartments.
Assumes that "refLocationProfilesAcup" has been previously defined
}
\examples{
 # See Vignette 4 and Vignette 5 for a full explanation
}
