% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixturePlotPanel.R
\name{mixturePlotPanel}
\alias{mixturePlotPanel}
\title{Panel of mixture plots of all pairwise simulated combinations}
\usage{
mixturePlotPanel(
  refLocationProfilesAcup,
  totProt,
  NstartMaterialFractions,
  errorReturn = FALSE,
  fitType,
  log2Transf = FALSE,
  eps = 0.001
)
}
\arguments{
\item{refLocationProfilesAcup}{data frame containing the Acup 
profiles for the reference compartments}

\item{totProt}{vector of total protein amounts 
(derived from a given amount of starting material) 
in each of the fractions comprising the profile}

\item{NstartMaterialFractions}{Number of fractions that reconstitute 
the starting material, e.g., a complete set of differential 
centrifugation fractions.  For experiment AT5, it is 6 
(N, M, L1, L2, P, and S).}

\item{errorReturn}{return all area-based errors if true}

\item{fitType}{use RSA, NSA, or Acup}

\item{log2Transf}{use log2-transformed values. Default is FALSE}

\item{eps}{small value to add so that log argument is greater than zero}
}
\value{
Panel of plots and a list of errors for all mixtures
}
\description{
For all sets of simulated proteins distributed between two compartments, 
  plot the CPA estimates for a given type of profile versus true 
  mixture proportions; assumes eight or fewer compartments 
  for proper formatting of the plots.
}
\examples{
data(refLocProfAcup)
data(totProtAT5)
mixturePlotPanel(refLocationProfilesAcup=
                   refLocProfAcup, 
                 totProt=totProtAT5, NstartMaterialFractions=6, 
                 errorReturn = TRUE, 
                 fitType="RSA", log2Transf=FALSE) 
# It may be necessary to open a separate window to
# properly display this panel of plots. Further details are in Tutorial 4.
}
