% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protDistanceCompute.R
\name{nearestProts}
\alias{nearestProts}
\title{Compute distances between a particular protein or profile
 and all other proteins,
 and list the nearest ones}
\usage{
nearestProts(protName, n.nearest = 5, distProts, protNames, profile)
}
\arguments{
\item{protName}{Name of protein to which distances are to be computed}

\item{n.nearest}{Number of nearest proteins to list}

\item{distProts}{distance matrix created by, for example}

\item{protNames}{A list of all proteins in a dataset}

\item{profile}{dataframe of profiles for proteins}
}
\value{
List of the proteins in protName closest to protName or to profile
}
\description{
Compute distances between a particular protein or profile
 and all other proteins,
 and list the nearest ones
}
\examples{
data(protNSA_test)
distUse <- dist(protNSA_test[,seq_len(9)], method='euclidean')
protsUse <- names(protNSA_test)
nearestProts(protName='CTSD', n.nearest=10,  distProts=distUse,
  protNames=protsUse, profile=protNSA_test[,seq_len(9)])
}
