% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAtransform.r
\name{NSAfromAcup}
\alias{NSAfromAcup}
\title{Compute relative specific amount from
 relative amounts in protein fractions}
\usage{
NSAfromAcup(Acup, NstartMaterialFractions, totProt = NULL)
}
\arguments{
\item{Acup}{amount of given protein in fraction /
amount of that given protein in starting material}

\item{NstartMaterialFractions}{Number of differential fractions,
typically 6, for N, M, L1, L2, P, and S}

\item{totProt}{Total protein counts in each of the differential
and nycodenz fractions; this is necessary to compute RSA's}
}
\value{
NSA: normalized specific amount
}
\description{
Compute relative specific amount from
 relative amounts in protein fractions
}
\examples{
data(protNSA_test)
data(totProtAT5)
protAcup <- AcupFromNSA(protNSA_test[,seq_len(9)],
            NstartMaterialFractions = 9, totProt=totProtAT5)
protNSA <- NSAfromAcup(protAcup, NstartMaterialFractions = 9,
            totProt=totProtAT5)
all.equal(protNSA, protNSA_test[,seq_len(9)])
}
