% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAtransform.r
\name{NSAfromAcup}
\alias{NSAfromAcup}
\title{Convert Acup to NSA}
\usage{
NSAfromAcup(Acup, NstartMaterialFractions, totProt = NULL)
}
\arguments{
\item{Acup}{profiles;
amount of given protein in fraction /
amount of that given protein in starting material
(see help file for protAcup_test).}

\item{NstartMaterialFractions}{Number of fractions that reconstitute 
the starting material, e.g., a complete set of differential 
centrifugation fractions.  For experiment AT5, it is 6 
( N, M, L1, L2, P, and S).}

\item{totProt}{Number of fractions that reconstitute 
the starting material, e.g., a complete set of differential 
centrifugation fractions.  For experiment AT5, it is 6 
( N, M, L1, L2, P, and S).}
}
\value{
NSA A matrix giving normalized specific amount profiles 
   (see help file for protNSA_test)
}
\description{
Convert Acup (relative amount) profiles to NSA
 (normalized specific amount) profiles
}
\examples{
data(protAcup_test)
data(totProtAT5)
protNSA_out1 <- NSAfromAcup(protAcup_test[,seq_len(9)], 
            NstartMaterialFractions = 6,
            totProt=totProtAT5)
round(head(protNSA_out1), digits=4)
}
