% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAtransform.r
\name{NSAfromRSA}
\alias{NSAfromRSA}
\title{Convert RSA to NSA}
\usage{
NSAfromRSA(RSA)
}
\arguments{
\item{RSA}{data frame containing RSA profiles.
RSA is the ratio of two ratios: 
the numerator is the amount of a given protein in a particular 
fraction divided by the amount of that given protein in the 
starting material while the denominator is amount of total protein 
in a particular fraction divided by the amount of total protein 
in the starting material. The RSA describes the fold-enrichment 
(RSA>1) or depletion (RSA<1) of a protein during the 
fractionation process, and is analogous to the 
relative specific activity term used in classical 
analytical subcellular fractionation.}
}
\value{
A data frame of NSA profiles 
   (see help file for protNSA_test)
}
\description{
Convert RSA (relative specific amount) profiles to NSA
(normalized specific amount) profiles
}
\examples{
data(protRSA_test)
protNSA_out2 <- NSAfromRSA(protRSA_test[,seq_len(9)])
round(head(protNSA_out2), digits=4)
str(protNSA_out2)
}
