% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierFind.R
\name{outlierFind}
\alias{outlierFind}
\title{Identify outlier spectra in a series of proteins}
\usage{
outlierFind(
  protClass,
  outlierLevel = "peptide",
  numRefCols = 5,
  numDataCols = 9,
  outlierMeth = "boxplot",
  range = 3,
  proba = 0.99,
  eps = eps,
  randomError = TRUE,
  setSeed = NULL,
  cpus = 1
)
}
\arguments{
\item{protClass}{a matrix of protein, peptide identifiers and
normalized specific amounts}

\item{outlierLevel}{peptide for outlier spectra within peptides, or
protein for outlier peptides within proteins`}

\item{numRefCols}{number of columns before Mass Spectrometry
data columns}

\item{numDataCols}{how many columns in MS data}

\item{outlierMeth}{boxplot (recommended), scores, or none}

\item{range}{the range parameter used for identifying outliers}

\item{proba}{probability to exclude outler for scores method}

\item{eps}{value to add before log2 transfromations
(to avoid taking log of zero)}

\item{randomError}{TRUE if allow it to be random}

\item{setSeed}{seed for random number generator}

\item{cpus}{1 (default);
if cpus > 1 use BiocParallel with SnowParm(cpus)}
}
\value{
additional column of indicators of outlier peptides or outlier
           spectra for a set of proteins or peptides
}
\description{
Identify outlier spectra profiles for a series of proteins or 
  protein identifiers using both scoresMod and boxplotMod
}
\examples{
set.seed(17356)
eps <- 0.029885209
data(spectraNSA_test)
flagSpectraBox <- outlierFind(protClass=spectraNSA_test,
                              outlierLevel='peptide', numRefCols=5,
                              numDataCols=9,
                              outlierMeth='boxplot', range=3, eps=eps,
                              randomError=TRUE, cpus=2)
# examine numbers of spectra that are outliers
table(flagSpectraBox$outlier.num.spectra)
}
