% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierFind.R
\name{outlierFind_i}
\alias{outlierFind_i}
\title{Find outlier spectra within a protein}
\usage{
outlierFind_i(
  i,
  protClass,
  outlierLevel = "peptide",
  numRefCols,
  numDataCols,
  outlierMeth,
  range,
  proba,
  eps = eps,
  randomError
)
}
\arguments{
\item{i}{unique identifier for protein i}

\item{protClass}{a matrix of protein, peptide identifiers and
normalized specific amounts}

\item{outlierLevel}{peptide for outlier spectra within peptides, or
protein for outlier peptides within proteins`}

\item{numRefCols}{number of columns before Mass
Spectrometry data columns}

\item{numDataCols}{how many columns in MS data}

\item{outlierMeth}{boxplot (recommended), scores, or none}

\item{range}{the range parameter used for identifying outliers}

\item{proba}{probability to exclude
outlier for scores method}

\item{eps}{value to add before log2 transfromations (to avoid
taking log of zero)}

\item{randomError}{TRUE if allow it to be random}
}
\value{
additional indicator of outlier peptides or outlier
           spectra for a set of proteins or peptides
}
\description{
Service function for outlierFind to identify
outlier spectra within a protein using both scoresMod and
boxplotMod
}
\examples{
set.seed(63561)
eps <- 0.029885209
data(spectraNSA_test)
uniqueLabel <- spectraNSA_test$pepId[1]
flagSpectraBox <- outlierFind_i(i=uniqueLabel, protClass=TLN1_test,
                      outlierLevel='peptide', numRefCols=5, numDataCols=9,
                      outlierMeth='boxplot', range=3, proba=NA, eps=eps,
                      randomError=TRUE)
str(flagSpectraBox)
# examine numbers of spectra that are outliers
}
