% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileSummarize.r
\name{profileSummarize}
\alias{profileSummarize}
\title{Calculate mean protein or peptide profiles and standard errors}
\usage{
profileSummarize(
  protsCombineCnew,
  numRefCols,
  numDataCols,
  refColsKeep = c(1, 2),
  eps,
  GroupBy = "protId",
  outlierExclude = "spectraAndPeptide",
  cpus = 1,
  singularList = FALSE
)
}
\arguments{
\item{protsCombineCnew}{a data frame containing profiles 
with outlier information}

\item{numRefCols}{number of fractions in each profile}

\item{numDataCols}{how many columns in MS data}

\item{refColsKeep}{which reference (non-data) columns to keep}

\item{eps}{epsilon to avoid log(0)}

\item{GroupBy}{'protId' if average by protein; 'peptideId'
if average by peptide}

\item{outlierExclude}{'none', 'spectra', or
'spectraAndpeptide' (default)
 according to exclusion level}

\item{cpus}{1 (default);
if cpus > 1 use BiocParallel with SnowParm(cpus)}

\item{singularList}{if TRUE, list proteins and channels with singular fit 
from the random effects function lmer; default is FALSE}
}
\value{
Mean or weighted mean NSA profiles
}
\description{
ProfileSummarize calculates mean and SE for each channel in each 
    protein or peptide profile using a random effect model (proteins)
    or arithmetic mean (peptides).  A random effect model can avoid 
    dominance of a peptide with a large number spectra.  
    See Tutorial 6 for details.
}
\examples{
set.seed(17356)
eps <- 0.029885209
data(spectraNSA_test)
flagSpectraBox <- outlierFind(protClass=spectraNSA_test,
           outlierLevel='peptide', numRefCols=5, numDataCols=9,
           outlierMeth='boxplot', range=3, eps=eps,
           randomError=TRUE, cpus=1)
# examine numbers of spectra that are outliers
table(flagSpectraBox$outlier.num.spectra)
pepProfiles <- profileSummarize(protsCombineCnew=flagSpectraBox,
           numRefCols=6, numDataCols=9, refColsKeep=c(1,2,4),eps=eps,
           GroupBy='peptideId', outlierExclude='spectra', cpus=1)
str(pepProfiles, strict.width='cut', width=65)

}
