% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileSummarize.r
\name{profileSummarize}
\alias{profileSummarize}
\title{protProfileSummarize calculates mean and SE for each channel in
each prot using random effect model or arithmetic mean
random effect model can avoid dominance of a sequence with
too many spectra}
\usage{
profileSummarize(
  protsCombineCnew,
  numRefCols,
  numDataCols,
  refColsKeep = c(1, 2),
  eps,
  GroupBy = "protId",
  outlierExclude = "spectraAndPeptide",
  cpus = 1,
  singularList = FALSE
)
}
\arguments{
\item{protsCombineCnew}{a matrix of protein information and normalized
specific amounts and outlier information}

\item{numRefCols}{number of columns before Mass Spectrometry
data columns}

\item{numDataCols}{how many columns in MS data}

\item{refColsKeep}{which reference (non-data) columns to keep}

\item{eps}{epsilon to avoid log(0)}

\item{GroupBy}{'protId' if average by protein; 'peptideId'
if average by peptide}

\item{outlierExclude}{'none', 'spectra', or
'spectraAndpeptide' (default)
 according to exclusion level}

\item{cpus}{1 (default);
if cpus > 1 use BiocParallel with SnowParm(cpus)}

\item{singularList}{If true, list proteins and channels with singular
lmer fit results; default is FALSE}
}
\value{
protProfileSummaryIdentity: mean or weighted mean normalized
         specific amount profiles
}
\description{
protProfileSummarize calculates mean and SE for each channel in
each prot using random effect model or arithmetic mean
random effect model can avoid dominance of a sequence with
too many spectra
}
\examples{
set.seed(17356)
eps <- 0.029885209
data(TLN1_test)
flagSpectraBox <- outlierFind(protClass=TLN1_test,
           outlierLevel='peptide', numRefCols=5, numDataCols=9,
           outlierMeth='boxplot', range=3, eps=eps,
           randomError=TRUE, cpus=1)
# examine numbers of spectra that are outliers
table(flagSpectraBox$outlier.num.spectra)
pepProfiles <- profileSummarize(protsCombineCnew=flagSpectraBox,
           numRefCols=6, numDataCols=9, refColsKeep=c(1,2,4),eps=eps,
           GroupBy='peptideId', outlierExclude='spectra', cpus=1)
str(pepProfiles, strict.width='cut', width=65)

}
