% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{protCap}
\alias{protCap}
\title{convert single protein name to capitalize first character only}
\usage{
protCap(x)
}
\arguments{
\item{x}{a protein name}
}
\value{
single protein name to capitalize first character only
}
\description{
convert single protein name to capitalize first character only
}
\examples{
protCap('tpp1')
}
