% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDataPrep.R
\name{proteinDataPrep}
\alias{proteinDataPrep}
\title{The program ensures that spectra are nested within peptides, and these
 are nested within proteins, and all peptide and protein data are
 contiguous. That is, not split up into separate locations
This program should be run first on a comma-separeted data file}
\usage{
proteinDataPrep(protClass, numRefCols, numDataCols)
}
\arguments{
\item{protClass}{frame of protein and peptide names and
data as described above}

\item{numRefCols}{number of reference columns}

\item{numDataCols}{numer of data columns}
}
\value{
data frame like the input, but with spectra and peptides
       nested within proteins in same location
}
\description{
The first column must contain prot name, and the second must be the peptide
The columns after 'numRefCols' contain relative abundance levels
There should be 'numDataCols' abundance level columns
}
\examples{
data(TLN1_test)
TLN1fixed <- proteinDataPrep(TLN1_test, numRefCols=5,numDataCols=9)
head(TLN1fixed)
}
