% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinMix.r
\name{proteinMix}
\alias{proteinMix}
\title{Simulate sets of protein profiles distributed between two compartments}
\usage{
proteinMix(AcupRef, Loc1, Loc2, increment.prop = 0.1)
}
\arguments{
\item{AcupRef}{data frame containing Acup profiles for the 
reference compartments}

\item{Loc1}{row number of one compartment}

\item{Loc2}{row number of other compartment}

\item{increment.prop}{fraction increment from 0 to 1}
}
\value{
A data frame containing Acup profiles for the simulated proteins
}
\description{
Compute the profile of simulated proteins that are distributed 
between two compartments in specified proportions
}
\examples{

data(refLocProfAcup)

# Compute mixtures
mixCytoLysoAcup <- proteinMix(AcupRef=refLocProfAcup,
                              increment.prop=0.1,
                              Loc1=1, Loc2=4)
}
