% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{protIndex}
\alias{protIndex}
\title{return index of a protein name, or (if exactMatch=TRUE) indices of
   proteins starting with the string given in 'protName'}
\usage{
protIndex(protName, profile, exactMatch = FALSE)
}
\arguments{
\item{protName}{name of protein to search for}

\item{profile}{data frame of protein identifiers (protName) and
their relative abundance in centrifugation fractions}

\item{exactMatch}{default is F}
}
\value{
the protein name and its index (row in profile)
}
\description{
return index of a protein name, or (if exactMatch=TRUE) indices of
   proteins starting with the string given in 'protName'
}
\examples{
data(protNSA_test)
protIndex('TLN1', profile=protNSA_test)
}
