% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protPepProfilePlot.R
\name{protPepPlotfun}
\alias{protPepPlotfun}
\title{Plot profile and peptide profiles with reference profiles}
\usage{
protPepPlotfun(
  protName,
  protProfile,
  Nspectra = TRUE,
  pepProfile = NULL,
  numRefCols,
  numDataCols,
  n.compartments = 8,
  refLocationProfiles,
  assignPropsMat,
  propCI = FALSE,
  transType = "",
  yAxisLabel = ""
)
}
\arguments{
\item{protName}{Name of the protein to plot}

\item{protProfile}{protein profile}

\item{Nspectra}{indicator for if there are columns in profile for Nspectra
(number of spectra) and Npep (number of peptides)}

\item{pepProfile}{peptide profiles}

\item{numRefCols}{number of reference columns
(preceding the data profile columns)}

\item{numDataCols}{number of fractions per protein}

\item{n.compartments}{number of compartments (8 in Jadot data)}

\item{refLocationProfiles}{A matrix refLocationProfiles giving the
  abundance level profiles of the subcellular locations
n.compartments = 8 columns are subcellular locations, and
numDataCols rows are the fraction names}

\item{assignPropsMat}{A matrix of assignment proportions, from the
  constrained proportional assignment algorithm,
and optionally upper and lower 95 percent confidence limits}

\item{propCI}{True if lower and upper confidence intervals
are included in assignPros}

\item{transType}{label for transformation; default is ""}

\item{yAxisLabel}{label for y-axis}
}
\value{
plot of average, peptide, and reference profiles
}
\description{
This function plots the average profiles of any protein in the dataset,
  the peptide profiles, and also the reference profile for each compartment
}
\examples{
  # See Vignette 6 for a full explanation
}
