% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protPepProfilePlot.R
\name{protPepPlotfun}
\alias{protPepPlotfun}
\title{Plot reference compartment profiles overlaid with the profile 
  of a single protein and its component peptides}
\usage{
protPepPlotfun(
  protName,
  protProfile,
  Nspectra = TRUE,
  pepProfile = NULL,
  numRefCols,
  numDataCols,
  n.compartments = 8,
  refLocationProfiles,
  assignPropsMat,
  propCI = FALSE,
  transType = "",
  yAxisLabel = ""
)
}
\arguments{
\item{protName}{Name of the protein to plot}

\item{protProfile}{Data frame containing protein profiles}

\item{Nspectra}{Data frame containing protein profile}

\item{pepProfile}{Data frame containing peptide profiles}

\item{numRefCols}{Number of reference columns
(preceding the data profile columns)}

\item{numDataCols}{Number of fractions in a profile}

\item{n.compartments}{Number of compartments}

\item{refLocationProfiles}{A data frame containing profiles of the 
subcellular locations}

\item{assignPropsMat}{A data frame containing CPA values}

\item{propCI}{TRUE if lower and upper confidence intervals are included 
in assignPropsMat (not currently implemented)}

\item{transType}{Label for y-axis on each individual plot (default none))}

\item{yAxisLabel}{Label for y-axis of entire panel}
}
\value{
plot of average, peptide, and reference profiles
}
\description{
This function creates a panel of plots, one for each compartment.  
     Each plot contains a reference compartment profile overlaid 
     with the profiles of one protein of interest and its component
      peptides as well as the CPA value for that compartment.
}
\examples{
  # See Tutorial 6 for a full explanation
}
