% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protProfilePlot.R
\name{protPlotfun}
\alias{protPlotfun}
\title{Plot profiles of reference proteins}
\usage{
protPlotfun(
  protName,
  profile,
  finalList = NULL,
  numDataCols = 9,
  n.compartments = 8,
  refLocationProfiles,
  assignPropsMat,
  transType = "",
  yAxisLabel = ""
)
}
\arguments{
\item{protName}{Name of the protein to plot}

\item{profile}{Data frame of protein names and their relative
abundance levels..}

\item{finalList}{spectrum-level abundance levels by protein and
peptide; Ehis is NULL
  if not available}

\item{numDataCols}{number of fractions per protein}

\item{n.compartments}{number of compartments (8 in Jadot data)}

\item{refLocationProfiles}{A matrix refLocationProfiles giving
the abundance level profiles of the subcellular locations
n.compartments = 8 columns are subcellular locations,
and numDataCols rows are the fraction names}

\item{assignPropsMat}{A matrix of assignment proportions
for proteins of interest, from the constrained proportional
assignment algorithm,
and optionally upper and lower 95 percent confidence limits.
 Can be a single protein}

\item{transType}{type of transformation (for plot label)}

\item{yAxisLabel}{label for y-axis if present}
}
\value{
plot of protein, peptide, and reference profiles
}
\description{
This function plots the average profiles of any protein in the dataset,
  the peptide profiles, and also the reference profile for each compartment
}
\examples{
data(protNSA_test)
data(markerListJadot)
refLocationProfilesNSA <- locationProfileSetup(profile=protNSA_test,
        markerList=markerListJadot, numDataCols=9)
protCPAfromNSA_test <- fitCPA(profile=protNSA_test,
                              refLocationProfiles=refLocationProfilesNSA,
                              numDataCols=9)
protPlotfun(protName="TLN1", profile=protNSA_test,
            numDataCols=9, n.compartments=8,
            refLocationProfiles=refLocationProfilesNSA,
            assignPropsMat=protCPAfromNSA_test,
            yAxisLabel="Normalized Specific Amount")
}
