% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{Qfun4}
\alias{Qfun4}
\title{goodness-of-fit
internal function; goodness-of-fit measure}
\usage{
Qfun4(pvec, y, gmat, methodQ = "sumsquares")
}
\arguments{
\item{pvec}{vector of assignment proportions to
compartments; to be estimated; must sum to 1}

\item{y}{observed data for amounts of protein in fractions;
must sum to 1}

\item{gmat}{reference protein amounts}

\item{methodQ}{either 'sumsquares' (default) or 'sumabsvalue'}
}
\value{
value of goodness-of-fit function
}
\description{
goodness-of-fit
internal function; goodness-of-fit measure
}
\examples{
# suppose there are four fractions and three compartments
yy <- c(0.1, 0.2, 0.1, 0.6)
gmat <- matrix(c(0,   0,   1,   # ref prot 1, all in compartment 3
                 0,   1,   0,   # ref prot 2, all in compartment 2
                 0,   0.5, 0.5, # ref prot 3, half in comp 2, half in 3)
                 0.5, 0,   0.5),# ref prot 4, half in comp 1, half in 3
                 ncol=3, byrow=TRUE )
pvec <- c(0, 0.3, 0.7)  # 30\%  in comp 2, 70\% in comp 3
Qfun4(pvec, yy, gmat)
}
