% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{Qfun4}
\alias{Qfun4}
\title{Goodness-of-fit internal function}
\usage{
Qfun4(pvec, y, gmat, methodQ = "sumsquares")
}
\arguments{
\item{pvec}{vector of values between 0 and 1 summing to one 
(length = number of compartments)}

\item{y}{specified profile}

\item{gmat}{matrix of reference profiles}

\item{methodQ}{either 'sumsquares' (default) or 'sumabsvalue'}
}
\value{
Value of goodness-of-fit function
}
\description{
An internal function used to assess the goodness-of-fit of a 
 weighted mixture of reference profiles to a specified profile
}
\examples{

# Suppose a profile consists of four fractions and there are three
#  reference compartment, designated A, B, and C

A <-c(1, 0, 0, 0)
B <- c(0, .5, 0.5,0)
C <-c(0, 0, 0, 1)
gmat<-cbind(A, B,C)
# make vector for a specified profile to be evaluated
yy <- c(0.2, 0.15, 0.15, 0.5)
# make vector for candidate CPA value to be evaluated
pvec1 <- c(1, 0, 0)  # 100\% in compartment A, 0\% in compartment B2,
 #  0\% in compartment C
pvec2<-c(0, 1, 0)  # 0\% in compartment A, 100\% in compartment B2, 
 #  0\% in compartment C
pvec3<-c(0,0,1)  # 0\% in compartment A, 0\% in compartment B2, 
 #  100\% in compartment C
pvec4<-c(0.2, 0.3, 0.5) #20\% in compartment A, 30\% in compartment B2, 
 #  50\% in compartment C

Qfun4(pvec1, yy, gmat)
Qfun4(pvec2, yy, gmat)
Qfun4(pvec3, yy, gmat)
Qfun4(pvec4, yy, gmat)

}
