% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{Qfun4subset}
\alias{Qfun4subset}
\title{constrained goodness-of-fit}
\usage{
Qfun4subset(
  pvec.vary,
  yy,
  gmat,
  methodQ = "sumsquares",
  ind.vary,
  ind.fixed,
  par.fixed
)
}
\arguments{
\item{pvec.vary}{vector of assignment proportions to
compartments; to be estimated}

\item{yy}{observed data for amounts of protein in
fractions}

\item{gmat}{reference protein amounts}

\item{methodQ}{either 'sumsquares' (default) or
'sumabsvalue'}

\item{ind.vary}{if not NULL, indexes of varying parameters}

\item{ind.fixed}{indexes of fixed parameters; all parameters
must be one or the other}

\item{par.fixed}{values of fixed parameters; typically 0}
}
\value{
value of goodness-of-fit function
}
\description{
internal function; goodness-of-fit measure for
subset of varying parameters while other parameters
are fixed at zero.
}
\examples{
# suppose there are four fractions and three compartments
yy <- c(0.1, 0.2, 0.1, 0.6)
gmat <- matrix(c(0,   0,   1,   # ref prot 1, all in compartment 3
                 0,   1,   0,   # ref prot 2, all in compartment 2
                 0,   0.5, 0.5, # ref prot 3, half in comp 2, half in 3)
                 0.5, 0,   0.5),# ref prot 4, half in comp 1, half in 3
                 ncol=3, byrow=TRUE )
pvec <- c(0, 0.3, 0.7)  # 30\%  in comp 2, 70\% in comp 3
pvec.vary <- c(.3, 0.7)  # just the compartments that vary
ind.vary <- c(2,3)  # compartments 2 and 3 may vary
ind.fixed <- 1   # fix value of compartment 1
par.fixed <- 0   # fix that value at 0
Qfun4subset(pvec.vary, yy, gmat, ind.vary=ind.vary,
           ind.fixed=ind.fixed, par.fixed=par.fixed)
# Now allow only compartments 1 and 3 to vary
Qfun4subset(pvec.vary, yy, gmat, ind.vary=c(1,3),
           ind.fixed=2, par.fixed=0)

}
