% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAtransform.r
\name{RSAfromAcup}
\alias{RSAfromAcup}
\title{Convert Acup to RSA}
\usage{
RSAfromAcup(Acup, NstartMaterialFractions, totProt = NULL)
}
\arguments{
\item{Acup}{data frame of specified protein (row name) Acup profiles;
amount of given protein in fraction /
amount of that given protein in starting material
(see help file for protAcup_test).}

\item{NstartMaterialFractions}{Number of fractions that reconstitute 
the starting material, e.g., a complete set of differential 
centrifugation fractions.  For experiment AT5, it is 6 
( N, M, L1, L2, P, and S).}

\item{totProt}{vector of total protein amounts (derived from a 
given amount of starting material) in each of the 
fractions comprising the profile}
}
\value{
RSA profiles.  RSA is the ratio of two ratios: 
   the numerator is the amount of a given protein in a particular 
   fraction divided by the amount of that given protein in the 
   starting material while the denominator is amount of total protein 
   in a particular fraction divided by the amount of total protein 
   in the starting material. The RSA describes the fold-enrichment 
   (RSA>1) or depletion (RSA<1) of a protein during the 
   fractionation process, and is analogous to the 
   relative specific activity term used in classical 
   analytical subcellular fractionation
}
\description{
Convert Acup (relative amount) profiles to RSA 
(relative specific amount) profiles
}
\examples{
data(protAcup_test)
data(totProtAT5)
protRSA_out1 <- RSAfromAcup(protAcup_test[,seq_len(9)], 
     NstartMaterialFractions = 6,
     totProt=totProtAT5)
round(head(protRSA_out1), digits=4)
str(protRSA_out1)
}
