% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaProgs.R
\name{vectorizeAll}
\alias{vectorizeAll}
\title{unitize series of vectors}
\usage{
vectorizeAll(protMatOrig)
}
\arguments{
\item{protMatOrig}{matrix of profiles}
}
\value{
matrix with all rows having unit length
}
\description{
normalize all rows of a matrix to have unit length
}
\examples{
data(protNSA_test)
vectorizeAll(protNSA_test[,1:9])
}
